--- a/math/mathcore/test/CMakeLists.txt
+++ b/math/mathcore/test/CMakeLists.txt
@@ -79,7 +79,8 @@ ROOT_ADD_GTEST(GradientUnit testGradient.cxx LIBRARIES Core MathCore Hist )
 
 ROOT_ADD_GTEST(GradientFittingUnit testGradientFitting.cxx LIBRARIES Core MathCore Hist)
 
-ROOT_ADD_GTEST(MulmodUnit mulmod.cxx)
+ROOT_ADD_GTEST(MulmodUnitOpt mulmod_opt.cxx)
+ROOT_ADD_GTEST(MulmodUnitNoInt128 mulmod_noint128.cxx)
 
 ROOT_ADD_GTEST(RanluxppEngineTests RanluxppEngine.cxx
         LIBRARIES Core MathCore)