--- a/math/mathcore/test/mulmod_noint128.cxx
+++ b/math/mathcore/test/mulmod_noint128.cxx
@@ -1,0 +1,13 @@
+// @(#)root/mathcore:$Id$
+// Author: Jonas Hahnfeld 11/2020
+
+/*************************************************************************
+ * Copyright (C) 1995-2020, Rene Brun and Fons Rademakers.               *
+ * All rights reserved.                                                  *
+ *                                                                       *
+ * For the licensing terms see $ROOTSYS/LICENSE.                         *
+ * For the list of contributors see $ROOTSYS/README/CREDITS.             *
+ *************************************************************************/
+
+#define ROOT_NO_INT128
+#include "mulmod.icc"