These changes reduce the average time per number from around 29ns to less than 10ns on a Ryzen 9 3900 using GCC 8.3.1. Using GCC 7.5.0, it now takes around 13ns per number instead of 17ns. With Clang 9.0.1 each random number now takes less than 9ns. For comparison, the implementation from https://github.com/sibidanov/ranluxpp/ (with assembly versions for the multiplication and modulo operation) also comes in at around 9ns.