@hahnjo 2020-11-19
@phsft-bot build just on ROOT-fedora30/cxx14, ROOT-debian10-i386/cxx14, mac1015/cxx17